refactor: rm /model
