+++ packages/examples/src/common/language-server-runner.ts
@@ -7,8 +7,8 @@ import { WebSocketServer } from 'ws';
 import { Server } from 'http';
 import express from 'express';
 import { getLocalDirectory } from '../utils/fs-utils.js';
-import { upgradeWsServer } from './server-commons.js';
-import { LanguageServerRunConfig } from '../model/server.js';
+import { LanguageServerRunConfig, upgradeWsServer } from './server-commons.js';
+
 /** LSP server runner */
 export const runLanguageServer = (
     languageServerRunConfig: LanguageServerRunConfig
