+++ packages/examples/src/groovy/server/main.ts
@@ -6,7 +6,8 @@
 import { runLanguageServer } from '../../common/language-server-runner.js';
 import { resolve } from 'path';
 import { groovyConfig } from '../config.js';
-import { LanguageName } from '../../model/shared.js';
+import { LanguageName } from '../../common/server-commons.js';
+
 export const runGroovyLanguageServer = (baseDir: string, relativeDir: string) => {
     const processRunPath = resolve(baseDir, relativeDir);
     runLanguageServer({
