--- packages/examples/src/model/index.ts
@@ -1,7 +1,0 @@
-/* --------------------------------------------------------------------------------------------
- * Copyright (c) 2024 TypeFox GmbH (http://www.typefox.io). All rights reserved.
- * Licensed under the MIT License. See License.txt in the project root for license information.
- * ------------------------------------------------------------------------------------------ */
-export * from './server.js';
-export * from './client.js';
-export * from './shared.js';
