--- packages/examples/src/model/server.ts
@@ -1,17 +1,0 @@
-/* --------------------------------------------------------------------------------------------
- * Copyright (c) 2024 TypeFox GmbH (http://www.typefox.io). All rights reserved.
- * Licensed under the MIT License. See License.txt in the project root for license information.
- * ------------------------------------------------------------------------------------------ */
-import { ServerOptions } from 'ws';
-import * as cp from 'child_process';
-import { LanguageName } from './shared.js';
-
-export interface LanguageServerRunConfig {
-    serverName: string;
-    pathName: string;
-    serverPort: number;
-    runCommand: LanguageName | string;
-    runCommandArgs: string[];
-    wsServerOptions: ServerOptions,
-    spawnOptions?: cp.SpawnOptions;
-}
