--- packages/examples/src/model/shared.ts
@@ -1,10 +1,0 @@
-/* --------------------------------------------------------------------------------------------
- * Copyright (c) 2024 TypeFox GmbH (http://www.typefox.io). All rights reserved.
- * Licensed under the MIT License. See License.txt in the project root for license information.
- * ------------------------------------------------------------------------------------------ */
-export enum LanguageName {
-    /** https://nodejs.org/api/cli.html  */
-    node = 'node',
-    /** https://docs.oracle.com/en/java/javase/21/docs/specs/man/java.html */
-    java = 'java'
-}
