+++ packages/examples/src/node.ts
@@ -9,4 +9,3 @@ export * from './json/server/json-server.js';
 export * from './json/server/main.js';
 export * from './python/server/main.js';
 export * from './utils/fs-utils.js';
-export * from './model/server.js';
